feat: replace group enter/kick with toggle API and use threads.enqueue
- Add toggle API for group member management (enter if not in group, kick if already in group)
- Move group business logic from GroupController to GroupService
- Replace threads.current()->dispatch with threads.enqueue in builtin functions to prevent crashes
- Rename variables from kicker to actor for consistency
- Update Lua script to use group:toggle instead of create_group for member management
- Add add_member and remove_member methods to group class
- Remove update method from group class
