+++ include/fb/game/server.h
@@ -176,9 +176,8 @@ public:
     [[nodiscard]] async::task<bool>                         create_group(character& me, const std::string& target_name);
     [[nodiscard]] async::task<void>                         destroy_group(character& me);
     [[nodiscard]] async::task<bool>                         handle_group_action(character& actor, const std::string& target_name);
-    [[nodiscard]] async::task<void>                         enter_group_member(character& inviter, const std::string& target_name);
+    [[nodiscard]] async::task<void>                         toggle_group_member(character& actor, const std::string& target_name);
     [[nodiscard]] async::task<void>                         leave_group_member(character& leaver);
-    [[nodiscard]] async::task<void>                         kick_group_member(character& kicker, const std::string& target_name);
     [[nodiscard]] async::task<void>                         broadcast_group(uint32_t group_id, const std::string& message, MESSAGE_TYPE type);
     [[nodiscard]] async::task<void>                         create_clan(character& me, std::string name);
     [[nodiscard]] async::task<void>                         destroy_clan(character& me);
